HZN-1207: Split classification in engine and persistence modules
