+++ container/karaf/src/main/filtered-resources/etc/custom.properties
@@ -685,6 +685,7 @@ org.osgi.framework.system.packages.extra=org.apache.karaf.branding,\
         org.opennms.netmgt.config.users;version=${opennms.osgi.version},\
         org.opennms.netmgt.config.viewsdisplay;version=${opennms.osgi.version},\
         org.opennms.netmgt.config.webuiColors;version=${opennms.osgi.version},\
+        org.opennms.netmgt.flows.classification.persistence.api;version=${opennms.osgi.version},\
         org.opennms.api.integration.ticketing;version=${opennms.osgi.version},\
         org.opennms.api.reporting;version=${opennms.osgi.version},\
         org.opennms.api.reporting.parameter;version=${opennms.osgi.version},\
