+++ features/flows/classification/persistence/api/src/main/java/org/opennms/netmgt/flows/classification/persistence/api/Protocol.java
@@ -1,0 +1,86 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2017-2017 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.netmgt.flows.classification.persistence.api;
+
+import java.util.Objects;
+
+import com.google.common.base.Preconditions;
+
+public class Protocol {
+    private int decimal;
+    private String keyword;
+    private String description;
+
+    public Protocol(int decimal, String keyword, String description) {
+        Objects.requireNonNull(keyword);
+        Preconditions.checkArgument(decimal >= 0, "decimal must be >= 0");
+        this.decimal = decimal;
+        this.keyword = keyword;
+        this.description = description;
+    }
+
+    public int getDecimal() {
+        return decimal;
+    }
+
+    public void setDecimal(int decimal) {
+        this.decimal = decimal;
+    }
+
+    public String getKeyword() {
+        return keyword;
+    }
+
+    public void setKeyword(String keyword) {
+        this.keyword = keyword;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Protocol protocol = (Protocol) o;
+        return Objects.equals(decimal, protocol.decimal)
+                && Objects.equals(keyword, protocol.keyword)
+                && Objects.equals(description, protocol.description);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(decimal, keyword, description);
+    }
+}
