+++ features/flows/classification/persistence/impl/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <parent>
+      <groupId>org.opennms.features.flows.classification</groupId>
+      <artifactId>org.opennms.features.flows.classification.persistence</artifactId>
+      <version>22.0.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.opennms.features.flows.classification.persistence</groupId>
+  <artifactId>org.opennms.features.flows.classification.persistence.impl</artifactId>
+  <name>OpenNMS :: Features :: Flows :: Classification :: Persistence :: Impl.</name>
+  <packaging>bundle</packaging>
+  <dependencies>
+      <dependency>
+          <groupId>org.opennms.features.flows.classification.persistence</groupId>
+          <artifactId>org.opennms.features.flows.classification.persistence.api</artifactId>
+          <version>${project.version}</version>
+      </dependency>
+  </dependencies>
+</project>
