+++ features/flows/classification/persistence/impl/src/main/resources/META-INF/opennms/component-dao.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns:context="http://www.springframework.org/schema/context"
+	xmlns:tx="http://www.springframework.org/schema/tx" 
+	xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
+	xsi:schemaLocation="
+		http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
+		http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-4.2.xsd
+		http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
+	">
+
+  <context:annotation-config />
+  <tx:annotation-driven />
+
+  <bean id="classificationDao" class="org.opennms.netmgt.flows.classification.persistence.impl.StaticRuleClassificationDAO">
+  </bean>
+  <onmsgi:service interface="org.opennms.netmgt.flows.classification.persistence.api.ClassificationRuleDao" ref="classificationDao" />
+
+</beans>
