+++ features/flows/classification/persistence/pom.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <parent>
+      <groupId>org.opennms.features.flows</groupId>
+      <artifactId>org.opennms.features.flows.classification</artifactId>
+      <version>22.0.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.opennms.features.flows.classification</groupId>
+  <artifactId>org.opennms.features.flows.classification.persistence</artifactId>
+  <name>OpenNMS :: Features :: Flows :: Classification :: Persistence (Parent)</name>
+  <packaging>pom</packaging>
+  <modules>
+    <module>api</module>
+    <module>impl</module>
+  </modules>
+</project>
