+++ features/telemetry/adapters/netflow/pom.xml
@@ -51,8 +51,8 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
-      <groupId>org.opennms.features.flows</groupId>
-      <artifactId>org.opennms.features.flows.classification</artifactId>
+      <groupId>org.opennms.features.flows.classification</groupId>
+      <artifactId>org.opennms.features.flows.classification.engine</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
