+++ opennms-base-assembly/pom.xml
@@ -695,6 +695,11 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <groupId>org.opennms.features.flows.classification.persistence</groupId>
+      <artifactId>org.opennms.features.flows.classification.persistence.impl</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>org.opennms.features</groupId>
       <artifactId>org.opennms.features.newts</artifactId>
       <version>${project.version}</version>
